+++ lib/subghz/protocols/protocol_items.h
@@ -58,3 +58,4 @@
 #include "jarolift.h"
 #include "ditec_gol4.h"
 #include "keyfinder.h"
+#include "nord_ice.h"